fix(kaizen): #1720 Wave-1b — co-locate InvalidApiKeyOverride in client.py (cross-SDK error-taxonomy parity)
CI cross_sdk_parity test_no_python_only_errors_leak_public_surface flagged
InvalidApiKeyOverride as a Python-only public error in kaizen.llm.errors not
present in the Rust taxonomy. Move it to client.py beside its sibling
UnsupportedApiKeyOverride — both are client-layer BYOK-override guards
subclassing AuthError, NOT part of the cross-SDK-mirrored errors taxonomy
(UnsupportedApiKeyOverride already lives in client.py and passed). Behavior
unchanged; callers still catch AuthError. (Local tests/unit/llm scope missed
this — tests/cross_sdk_parity/ now included in the verify scope.)
